chore(create-elm-app): Move create-elm-app to /packages
