+++ packages/create-elm-app/README.md
@@ -1,0 +1,7 @@
+# create-elm-app
+
+This package includes scripts and configuration used by [Create Elm App](https://github.com/halfzebra/create-elm-app).<br>
+Please refer to its documentation:
+
+* [Getting Started](https://github.com/halfzebra/create-elm-app/blob/master/README.md#getting-started) – How to create a new app.
+* [User Guide](https://github.com/halfzebra/create-elm-app/blob/master/packages/craete-elm-app/template/README.md) – How to develop apps bootstrapped with Create React App.
