+++ packages/create-elm-app/config/paths.js
