+++ packages/create-elm-app/config/webpackDevServer.config.js
