+++ packages/create-elm-app/package.json
@@ -1,0 +1,95 @@
+{
+  "name": "create-elm-app",
+  "description": "Elm apps with zero configuration",
+  "version": "1.9.1",
+  "author": "Eduard Kyvenko <eduard.kyvenko@gmail.com>",
+  "bin": {
+    "elm-app": "bin/elm-app-cli.js",
+    "create-elm-app": "bin/create-elm-app-cli.js"
+  },
+  "bugs": {
+    "url": "https://github.com/halfzebra/create-elm-app/issues",
+    "email": "eduard.kyvenko@gmail.com"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "@babel/cli": "7.0.0-beta.56",
+    "@babel/core": "7.0.0-beta.56",
+    "@babel/plugin-transform-runtime": "7.0.0-beta.56",
+    "@babel/preset-env": "7.0.0-beta.56",
+    "assets-webpack-plugin": "^3.5.1",
+    "autoprefixer": "^8.0.0",
+    "babel-loader": "8.0.0-beta.4",
+    "babel-runtime": "^6.26.0",
+    "chalk": "^2.3.1",
+    "clean-webpack-plugin": "^0.1.18",
+    "cli-table": "0.3.1",
+    "connect-history-api-fallback": "^1.5.0",
+    "cross-spawn": "^6.0.4",
+    "css-loader": "^0.28.9",
+    "dotenv": "^5.0.0",
+    "elm": "^0.18.0",
+    "elm-hot-loader": "0.5.4",
+    "elm-test": "^0.18.12",
+    "elm-webpack-loader": "^4.4.0",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.6",
+    "fs-extra": "^5.0.0",
+    "html-webpack-plugin": "^2.30.1",
+    "http-proxy-middleware": "^0.17.4",
+    "minimist": "1.2.0",
+    "postcss-loader": "2.1.0",
+    "prompt": "1.0.0",
+    "react-dev-utils": "^5.0.0",
+    "react-error-overlay": "^4.0.0",
+    "string-replace-loader": "^1.3.0",
+    "style-loader": "^0.20.2",
+    "sw-precache-webpack-plugin": "^0.11.4",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.11.0",
+    "webpack-dev-server": "^2.11.1"
+  },
+  "devDependencies": {
+    "dir-compare": "^1.4.0",
+    "eslint": "^4.18.0",
+    "eslint-plugin-prettier": "^2.6.0",
+    "mocha": "^5.0.1",
+    "nightmare": "^2.10.0",
+    "rimraf": "^2.6.2",
+    "semantic-release": "^15.9.8",
+    "semantic-release-monorepo": "^6.0.3",
+    "unexpected": "^10.37.2"
+  },
+  "engines": {
+    "node": ">=6.0.0"
+  },
+  "files": [
+    "bin",
+    "config",
+    "scripts",
+    "template",
+    "README.md"
+  ],
+  "keywords": [
+    "cli",
+    "elm",
+    "webpack"
+  ],
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/halfzebra/create-elm-app"
+  },
+  "scripts": {
+    "eslint": "eslint .",
+    "prettier": "prettier --write --single-quote '{bin,config,scripts,template/src,tests}/**/*.js'",
+    "semantic-release": "semantic-release",
+    "test": "npm run test:functional",
+    "test:cli": "mocha tests/cliAccessibility.js --reporter spec --timeout 15000",
+    "test:functional": "mocha tests/*.spec.js --reporter spec --timeout 15000",
+    "build": "cd template && node ../bin/elm-app-cli.js build",
+    "start": "cd template && node ../bin/elm-app-cli.js start"
+  }
+}
