+++ packages/create-elm-app/scripts/build.js
