+++ packages/create-elm-app/scripts/utils/formatElmCompilerErrors.js
