+++ packages/create-elm-app/template/elm-package.json
