+++ packages/create-elm-app/template/gitignore
