+++ packages/create-elm-app/template/public/favicon.ico
