+++ packages/create-elm-app/template/public/index.html
