+++ packages/create-elm-app/template/public/manifest.json
