+++ packages/create-elm-app/template/src/Main.elm
