+++ packages/create-elm-app/template/src/index.js
