+++ packages/create-elm-app/template/src/registerServiceWorker.js
