+++ packages/create-elm-app/template/tests/elm-package.json
