+++ packages/create-elm-app/tests/cliAccessibility.js
